Add mark registration
The introduced function allows the user to register a new mark.
The user can insert a vote by sending a message to the bot. The message
must have a format like:
/voto IP Networks Design 30L 9 2020-12-25
This message requires the bot to insert the mark 30 cum laude for the
exam "IP Networks Design", that is worth 9 University credits and has
been taken on December 25th, 2020