Implement mark listing and deletion
Note that I inserted a student_exam_id column in the marks table.
This field is used by the users to specify the exam they want to delete
from the database. Since we don't want the user to see missing numbers
after having deleted an exam, the numbering must be rearranged. In order
to keep the insertion order of the marks, I had to decrement the
student_exam_id to all marks the user inserted
after the one that is being deleted.